share one REPL instance for REPL test cases
